fix: params of functions now match the types of the manual

+++ src/ft_bzero.c
@@ -1,4 +1,6 @@
-void	ft_bzero(void *buffer, unsigned long len)
+#include <stdio.h>
+
+void	ft_bzero(void *buffer, size_t len)
 {
 	unsigned char	*p;
